rework statistics endpoints
